Deep Learning/requirements.txt
async_timeout==3.0.1
brotli==1.1.0
brotlicffi==1.1.0.0
ca_certs_locater==1.0
cffi==1.17.1
chardet==3.0.4
contextlib2==21.6.0
cryptography==45.0.3
curio==1.6
cx_Oracle==8.3.0
cython==3.1.1
datasets==3.6.0
deep_translator==1.11.4
defusedxml==0.7.1
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
grandalf==0.8
grpc_reflection==1.0.0
grpc_tools==1.0.0
gspread==6.2.1
h2==4.2.0
hypothesis==6.131.28
importlib_metadata==8.6.1
importlib_metadata==8.0.0
importlib_resources==6.5.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.6
js==1.0
jsonschema==4.24.0
langchain_anthropic==0.3.13
langchain_astradb==0.6.0
langchain_aws==0.2.23
langchain_azure_ai==0.1.3
langchain_chroma==0.2.4
langchain_cohere==0.4.4
langchain_community==0.3.24
langchain_deepseek==0.1.3
langchain_elasticsearch==0.3.2
langchain_experimental==0.3.4
langchain_fireworks==0.3.0
langchain_google_vertexai==2.0.24
langchain_groq==0.3.2
langchain_huggingface==0.2.0
langchain_ibm==0.3.11
langchain_mistralai==0.2.10
langchain_mongodb==0.6.2
langchain_neo4j==0.4.0
langchain_nvidia_ai_endpoints==0.3.10
langchain_ollama==0.3.3
langchain_openai==0.3.18
langchain_perplexity==0.1.1
langchain_pinecone==0.2.6
langchain_postgres==0.0.14
langchain_qdrant==0.2.0
langchain_together==0.3.0
langchain_weaviate==0.0.4
langchain_xai==0.2.3
langchainhub==0.1.21
langsmith_pyo3==0.1.0rc5
lark==1.2.2
lxml==5.3.1
matplotlib==3.10.3
numexpr==2.10.2
numpy==2.2.6
oauth2client==4.1.3
objgraph==3.6.2
openapi_pydantic==0.5.1
outcome==1.3.0.post0
Pillow==11.2.1
psutil==7.0.0
Pygments==2.19.1
pygraphviz==1.14
pyodide==0.0.2
pyOpenSSL==25.1.0
pyppeteer==2.0.0
pysqlcipher3==1.2.0
pytest==8.3.5
python-dotenv==1.1.0
pyu2f==0.1.5
railroad==0.5.0
rapidfuzz==3.13.0
seaborn==0.13.2
setuptools_scm==8.3.1
simplejson==3.20.1
simsimd==6.2.1
slack_sdk==3.35.0
socksio==1.0.0
sqlcipher3==0.5.4
tensorflow==2.18.0
tensorflow_intel==2.18.0
tiktoken==0.9.0
tomli==2.2.1
tornado==6.4.2
uvicorn==0.34.2
uvloop==0.21.0
